+++ nab-tests/src/test/java/ru/hh/jclient/common/JClientTransactionTest.java
@@ -27,14 +27,14 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import ru.hh.jclient.common.util.storage.SingletonStorage;
-import ru.hh.nab.hibernate.HibernateTestConfig;
 import ru.hh.nab.hibernate.transaction.TransactionalScope;
 import ru.hh.nab.jclient.NabJClientConfig;
 import ru.hh.nab.jclient.checks.TransactionalCheck;
+import ru.hh.nab.jpa.JpaTestConfig;
 
 @ExtendWith(SpringExtension.class)
 @ContextConfiguration(
-    classes = {HibernateTestConfig.class, NabJClientConfig.class, JClientTransactionTest.TestConfig.class}
+    classes = {JpaTestConfig.class, NabJClientConfig.class, JClientTransactionTest.TestConfig.class}
 )
 public class JClientTransactionTest {
   private static final TestRequestDebug DEBUG = new TestRequestDebug(true);
