+++ nab-tests/src/test/java/ru/hh/nab/jpa/MappingConfigTest.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.hibernate;
+package ru.hh.nab.jpa;
 
 import jakarta.inject.Inject;
 import jakarta.persistence.EntityManager;
@@ -7,14 +7,13 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 import org.springframework.test.context.ContextConfiguration;
-import static ru.hh.nab.hibernate.HibernateTestConfig.TEST_PACKAGE;
-import ru.hh.nab.hibernate.model.PackageNotScanEntity;
-import ru.hh.nab.hibernate.model.TestEntity;
-import ru.hh.nab.hibernate.model.test.PackageScanEntity;
-import ru.hh.nab.jpa.MappingConfig;
+import static ru.hh.nab.jpa.JpaTestConfig.TEST_PACKAGE;
+import ru.hh.nab.jpa.model.PackageNotScanEntity;
+import ru.hh.nab.jpa.model.TestEntity;
+import ru.hh.nab.jpa.model.test.PackageScanEntity;
 import ru.hh.nab.testbase.jpa.JpaTestBase;
 
-@ContextConfiguration(classes = {HibernateTestConfig.class})
+@ContextConfiguration(classes = {JpaTestConfig.class})
 public class MappingConfigTest extends JpaTestBase {
 
   @Inject
@@ -37,7 +36,7 @@ public class MappingConfigTest extends JpaTestBase {
   }
 
   @Test
-  public void hibernateShouldHaveMappedEntities() {
+  public void persistenceProviderShouldHaveMappedEntities() {
     assertTrue(metamodelContainsEntity(entityManager, TestEntity.class));
     assertTrue(metamodelContainsEntity(entityManager, PackageScanEntity.class));
     assertFalse(metamodelContainsEntity(entityManager, PackageNotScanEntity.class));
