+++ nab-tests/src/test/java/ru/hh/nab/jpa/model/test/PackageScanEntity.java
@@ -1,4 +1,4 @@
-package ru.hh.nab.hibernate.model;
+package ru.hh.nab.jpa.model.test;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,7 +7,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 
 @Entity
-public class PackageNotScanEntity {
+public class PackageScanEntity {
   @Column(name = "id")
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Id
